requirements.txt
kafka-python
pyspark==3.5.0
streamlit
pandas
psycopg2-binary
requests